Completed development of class to summarize Asian males booked population changes.
